fix: stabilize admin login rendering
