+++ .changeset/auth-brand-snapshot.md
@@ -1,0 +1,5 @@
+---
+"@questpie/admin": patch
+---
+
+Cache admin auth branding snapshots to avoid React update loops on login pages.
