+++ examples/tanstack-barbershop/src/components/layout/Header.tsx
@@ -133,7 +133,6 @@ export function Header({
 					{/* Locale Switcher */}
 					<DropdownMenu>
 						<DropdownMenuTrigger
-							nativeButton={false}
 							render={
 								<Button
 									variant="ghost"
@@ -287,7 +286,6 @@ export function Header({
 									</span>
 									<DropdownMenu>
 										<DropdownMenuTrigger
-											nativeButton={false}
 											render={<Button variant="outline" size="sm" />}
 										>
 											<Icon icon="ph:globe" className="mr-2 size-4" />
